test: move responses api tests to existing transformation test file
